LayoutTests/fast/mediastream/microphone-change-while-muted-expected.txt

CONSOLE MESSAGE: A MediaStreamTrack ended due to a capture failure


PASS Detection of missing capturing device should trigger capture to fail even if track is muted
